disable persp-mode
Persp-mode has a very annoying behaviour when a new frame is created to display
a given buffer. It will re-create the current perspective on the new frame,
along with the window configuration. THis results in a new frame on which I'll
always do C-x 1 .

So, since I haven't been using perspectives and don't really understand them,
keep the package out of reach for the time being