fix[zsh/emacs]: fix emacs font loading
When emacsclient is given an empty string as alternate-editor, it invokes `emacs
--daemon`, which does not create a frame until after the daemon startup.

As a result, the setting of the default font behaves awkwardly:

- By letting spacemacs manage it, it calls `(set-frame-font ...)`. At the moment
  it is called, there's not yet a frame (which may be the reason of previous
  bugs with font size).

  When using a font such as 'Fira Code Retina', the accentuated characters
  áéíóúãõ... appear to render correctly, but everything else uses some other
  font that is much thinner.

- By setting spacemacs-default-font to nil and using customize.el to
  `custom-set-faces ...` , it seems that it loads the correct font, but the
  weights of normal and accentuated letters are slightly different -- enough for
  you to perceive, but close enough for you to doubt yourself.

In either of the previous cases, after using the GUI font selector or by
customizing the default face, the font would be loaded correctly.

The necessary fix seems to be calling emacs without --daemon. This opens the
frame *before* loading user initialization scripts, then everything works
correctly (just as executing `emacs` directly from the cmdline)