feat[emacs] kzk/open-fun function to call system's open
In MacOS, `open` is used to invoke the default app for a given file. In
GNU/Linux desktop, it's `xdg-open`.

org-mode, in linux, defers to mailcap default handlers, which are a royal mess
in my experience. The `xdg-open-fun` was used as default org-mode
open-with-system to avoid having to deal with mailcap.

This change makes the function work correctly in Linux and MacOS