fix[emacs]: stop using venv from within emacs
auto setting venv changes the VIRTUAL_ENV variable for the whole emacs process,
not just for the processes launched from the current project.

This has several bizarre effects, such as pyright loading the right venv only if
we have loaded the right venv. Working with several projects is, therefore, not
viable.

This change disables spacemacs auto-switching of virtual envs on project load.
For pyright, use pyrightconfig.json to set `venv` to the name of the project's
virtual env (pyright does not look to `.python-version`)