chore[emacs]: fix DISPLAY being set to $WAYLAND_DISPLAY in server
Because cthulu, running `emacsclient --alternate-editor= -c` the $DISPLAY
variable would be set to `wayland-0` (the value of $WAYLAND_DISPLAY). This has a
severe impact when emacs calls other graphical processes that still use the X
protocol -- e.g. browse-url could not correctly call chrome, as chrome is still
a X11 application.

This happens despite the value of DISPLAY in `/proc/$(pidof
emacs)/environ` (i.e. `environ` would show `DISPLAY=:0` , but `(getenv
"DISPLAY)` in emacs would return `wayland-0`

This happens with a completly vanilla emacs, which indicates it's not a
spacemacs issue or one of my setup.

Advising `setenv` and `setenv-internal` to display a backtrace on updates to
DISPLAY did not result in any update to the variable -- maybe it's updated in C
code ...

We can workaround the issue by having emacsclient call --display=$DISPLAY -- but
in this case, it starts gtk with the X11 backend instead of the wayland native
one.

This kludge patches the value of DISPLAY after the first frame is created,
avoiding the problem