fix[emacs]: make sure helm-org uses my custom ordering function for helm-org-in-buffer-headings
helm-org-in-buffer-headings uses, by default, a relevance ordering instead of
buffer ordering, which I find confusing, and it has no customization for the
ordering.

Simply redefining `helm-org-indent-headings` does not work, as the
package uses a hash-quoted reference to the function. The solution is to do a
little load-order-ballet, redefining the function as early as possible after the
original function.