test: automtically find .python-version and make it buffer local instead of setting the environgment variable
By default, pyvenv will set the VIRTUALENV environment variable, which
globally configures emacs to launch all child processes uwing that env.

This sucks if you have multiple projects opened.

This was an attempt to use pyvenv functionality to auto-active the
.python-version python version whenever a buffer was visited.

Issue is: Pyvenv does not restore original environment variables when it
alternates to a buffer that does not have pyvenv-activate set