fix[emacs]: Avoid opening terminal emacs inside emacs vterm
Opening a vterm inside emacs and then running `git commit` that would try to
pop up a terminal emacs can lead to issues -- pressing `:` to open `evil-ex`
takes the editor to a rather unexpected state, from which exiting seems to be
a mission requiring a magician.

This hack makes the editors be `emacsclient` , effectivelly popping up the
target buffer over the vterm window. Not the same as editing inside the
terminal, but close enough.

This _could_ be done with `with-editor-export-editor` and
`with-editor-export-git-editor` added to `vterm-mode-hook` -- but that would
not set the `alias`, so I currently chose to do it via shell.