feat[emacs]: provide consult-descbinds alternative to embark-bindings
For some reason, spacemacs nested keybindings are in the form

    `("Files" . (keymap . ((key . command) ...)))

This fails `keymapp` , which checks if the value is a cons-cell and if its car
is 'keymap .

embark-bindings tries to recursively read the nested keymaps, but since the
keymap fails the `keymapp` check, it never descends, leaving only

    SPC f    Files

in the candidates -- which does not help us much.

Besides, the associated object is `("Files" . (keymap . (...))) , which embark
then tries to use in `funcall` -- of course, it fails miserably.

Besides+1, spacemacs contains _recursive keymaps_ (mindblown) -- SPC k SPC
goes back to spacemacs-default-map, as the emacs-lisp layer explicitly uses
bind-map to do so. Embark does not deal well with this.

I've used `helm-descbinds` as a base and created a consult-descbinds kind of
command which I use to override `describe-bindings` just like helm-descbind
does. This brings back which-key paging (which I like, because reasons) and
provides a completing-read interface on which-key 'C-h' command.