fix[emacs]: do not use aw-dispatch-always
ace-window with aw-dispatch-always and a single window does not show shortcuts! It simply reads a
key for an action (split, delete, etc.)

I find that a bit too hard do see in the echo area -- kind of easy to do
something unexpected. I prefer selecting the single window instead and move on
with life