feat: speed up pcase--make-docstring in emacs-30 in eldoc and marginalia
In emacs-30, the new version of pcase--make-docstring tries to find the
relative filename for each pcase declaration.

lsp-mode makes heavy use of pcase cases, which, when combined with the high
number of packages installed in spacemacs, makes pcase--make-docstring very
slow:

1. For each pcase-case
1.1 For each file in load-path
1.1.1. call relative-file-name

relative-file-name ends up taking a hell lot of time, probably because it was
called way too many times... emacs kind of hangs -- it's not only high CPU
usage, it hangs for several seconds.

This kludge brings back emacs-29 version of the function and advises
marginalia and eldoc to avoid calling the emacs-30 version.

This brings the changes in pcase--make-docstring without the performance
issues in marginalia and eldoc