fix[emacs]: Avoid sending commands to clients that do not support it
lsp-mode broadcast commands to all active servers, ignoring if the server
actually supports that command.

rust-lsp does not like to receive commands it does not know about and that
triggers a warning everytime.

With lsp-copilot, that results in rust receiving the post-suggestion-insert
and a warning popping up.

I've tried to propose that as part of the copilot PR, but that was ignored, so
for now I'm running with this patched `command-execute` function.